+++ packages/webview_flutter/webview_flutter_platform_interface/lib/src/platform_interface/webview_cookie_manager.dart
@@ -1,0 +1,24 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'package:webview_flutter_platform_interface/src/types/webview_cookie.dart';
+
+/// Interface for a platform implementation of a cookie manager.
+///
+/// This interface should only be extended, not implemented, as more methods may be added later.
+abstract class WebViewCookieManagerPlatform {
+  /// Clears all cookies for all [WebView] instances.
+  ///
+  /// Returns true if cookies were present before clearing, else false.
+  Future<bool> clearCookies() {
+    throw UnimplementedError(
+        'clearCookies is not implemented on the current platform');
+  }
+
+  /// Sets a cookie for all [WebView] instances.
+  Future<void> setCookie(WebViewCookie cookie) {
+    throw UnimplementedError(
+        'setCookie is not implemented on the current platform');
+  }
+}