+++ packages/webview_flutter/webview_flutter_platform_interface/lib/src/platform_interface/webview_platform.dart
@@ -6,6 +6,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/widgets.dart';
 import 'package:webview_flutter_platform_interface/src/platform_interface/javascript_channel_registry.dart';
+import 'package:webview_flutter_platform_interface/src/types/webview_cookie.dart';
 
 import '../types/types.dart';
 import 'webview_platform_callbacks_handler.dart';
@@ -59,6 +60,7 @@ abstract class WebViewPlatform {
   /// Clears all cookies for all [WebView] instances.
   ///
   /// Returns true if cookies were present before clearing, else false.
+  @Deprecated('Use `WebViewCookieManager.clearCookies` instead.')
   Future<bool> clearCookies() {
     throw UnimplementedError(
         "WebView clearCookies is not implemented on the current platform");