+++ packages/webview_flutter/webview_flutter_platform_interface/lib/src/types/types.dart
@@ -11,3 +11,4 @@ export 'web_resource_error.dart';
 export 'web_resource_error_type.dart';
 export 'web_settings.dart';
 export 'webview_request.dart';
+export 'webview_cookie.dart';