+++ packages/webview_flutter/webview_flutter_platform_interface/test/src/method_channel/webview_method_channel_test.dart
@@ -590,6 +590,26 @@ void main() {
         ],
       );
     });
+
+    test('setCookie', () async {
+      await MethodChannelWebViewPlatform.setCookie(
+          WebViewCookie(name: 'foo', value: 'bar', domain: 'flutter.dev'));
+
+      expect(
+        log,
+        <Matcher>[
+          isMethodCall(
+            'setCookie',
+            arguments: {
+              'name': 'foo',
+              'value': 'bar',
+              'domain': 'flutter.dev',
+              'path': '/',
+            },
+          ),
+        ],
+      );
+    });
   });
 }